+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/PParams.hs
@@ -64,10 +64,11 @@ import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.BaseTypes (
   EpochInterval (..),
   NonNegativeInterval,
-  Nonce (NeutralNonce),
+  Nonce (NeutralNonce, Nonce),
   StrictMaybe (..),
   UnitInterval,
   isSNothing,
+  parseAsRational,
  )
 import qualified Cardano.Ledger.BaseTypes as BT (ProtVer (..))
 import Cardano.Ledger.Binary (
@@ -137,7 +138,7 @@ import Data.Aeson as Aeson (
   (.!=),
   (.:),
  )
-import qualified Data.Aeson as Aeson (Value)
+import qualified Data.Aeson.Types as Aeson
 import Data.ByteString (ByteString)
 import qualified Data.ByteString as BS
 import Data.Coerce (coerce)
@@ -470,36 +471,42 @@ alonzoPParamsPairs ::
   PParamsHKD Identity (AlonzoEra c) ->
   [a]
 alonzoPParamsPairs pp =
-  uncurry (.=) <$> alonzoPParamsHKDPairs (Proxy @Identity) pp
+  uncurry (.=)
+    <$> alonzoPParamsHKDPairs (Proxy @Identity) pp
+      ++ ["minUTxOValue" .= Aeson.Null]
 
 instance FromJSON (AlonzoPParams Identity era) where
   parseJSON =
     Aeson.withObject "PParams" $ \obj ->
       AlonzoPParams
-        <$> obj .: "minFeeA"
-        <*> obj .: "minFeeB"
+        <$> obj .: "txFeePerByte"
+        <*> obj .: "txFeeFixed"
         <*> obj .: "maxBlockBodySize"
         <*> obj .: "maxTxSize"
         <*> obj .: "maxBlockHeaderSize"
-        <*> obj .: "keyDeposit"
-        <*> obj .: "poolDeposit"
-        <*> obj .: "eMax"
-        <*> obj .: "nOpt"
-        <*> obj .: "a0"
-        <*> obj .: "rho"
-        <*> obj .: "tau"
-        <*> obj .: "decentralisationParam"
-        <*> obj .: "extraEntropy"
+        <*> obj .: "stakeAddressDeposit"
+        <*> obj .: "stakePoolDeposit"
+        <*> obj .: "poolRetireMaxEpoch"
+        <*> obj .: "stakePoolTargetNum"
+        <*> parseAsRational (obj .: "poolPledgeInfluence")
+        <*> parseAsRational (obj .: "monetaryExpansion")
+        <*> parseAsRational (obj .: "treasuryCut")
+        <*> parseAsRational (obj .: "decentralization")
+        <*> (obj .: "extraPraosEntropy" >>= jsonToNonce)
         <*> obj .: "protocolVersion"
         <*> obj .: "minPoolCost" .!= mempty
-        <*> obj .: "lovelacePerUTxOWord"
-        <*> obj .: "costmdls"
-        <*> obj .: "prices"
-        <*> obj .: "maxTxExUnits"
-        <*> obj .: "maxBlockExUnits"
-        <*> obj .: "maxValSize"
+        <*> obj .: "utxoCostPerByte"
+        <*> obj .: "costModels"
+        <*> obj .: "executionUnitPrices"
+        <*> obj .: "maxTxExecutionUnits"
+        <*> obj .: "maxBlockExecutionUnits"
+        <*> obj .: "maxValueSize"
         <*> obj .: "collateralPercentage"
         <*> obj .: "maxCollateralInputs"
+    where
+      jsonToNonce :: Aeson.Value -> Aeson.Parser Nonce
+      jsonToNonce Aeson.Null = return NeutralNonce
+      jsonToNonce x = Nonce <$> parseJSON x
 
 newtype CoinPerWord = CoinPerWord {unCoinPerWord :: Coin}
   deriving stock (Eq, Ord)
@@ -732,7 +739,7 @@ alonzoPParamsHKDPairs px pp =
   alonzoCommonPParamsHKDPairs px pp
     ++ shelleyCommonPParamsHKDPairsV8 px pp
     ++ shelleyCommonPParamsHKDPairsV6 px pp
-    ++ [("lovelacePerUTxOWord", hkdMap px (toJSON @CoinPerWord) (pp ^. hkdCoinsPerUTxOWordL @_ @f))]
+    ++ [("utxoCostPerByte", hkdMap px (toJSON @CoinPerWord) (pp ^. hkdCoinsPerUTxOWordL @_ @f))]
 
 -- | These are the fields that are common across all eras starting with Alonzo.
 alonzoCommonPParamsHKDPairs ::
@@ -743,11 +750,11 @@ alonzoCommonPParamsHKDPairs ::
   [(Key, HKD f Aeson.Value)]
 alonzoCommonPParamsHKDPairs px pp =
   shelleyCommonPParamsHKDPairs px pp
-    ++ [ ("costmdls", hkdMap px (toJSON @CostModels) (pp ^. hkdCostModelsL @era @f))
-       , ("prices", hkdMap px (toJSON @Prices) (pp ^. hkdPricesL @era @f))
-       , ("maxTxExUnits", hkdMap px (toJSON @ExUnits) (pp ^. hkdMaxTxExUnitsL @era @f))
-       , ("maxBlockExUnits", hkdMap px (toJSON @ExUnits) (pp ^. hkdMaxBlockExUnitsL @era @f))
-       , ("maxValSize", hkdMap px (toJSON @Natural) (pp ^. hkdMaxValSizeL @era @f))
+    ++ [ ("costModels", hkdMap px (toJSON @CostModels) (pp ^. hkdCostModelsL @era @f))
+       , ("executionUnitPrices", hkdMap px (toJSON @Prices) (pp ^. hkdPricesL @era @f))
+       , ("maxTxExecutionUnits", hkdMap px (toJSON @ExUnits) (pp ^. hkdMaxTxExUnitsL @era @f))
+       , ("maxBlockExecutionUnits", hkdMap px (toJSON @ExUnits) (pp ^. hkdMaxBlockExUnitsL @era @f))
+       , ("maxValueSize", hkdMap px (toJSON @Natural) (pp ^. hkdMaxValSizeL @era @f))
        , ("collateralPercentage", hkdMap px (toJSON @Natural) (pp ^. hkdCollateralPercentageL @era @f))
        , ("maxCollateralInputs", hkdMap px (toJSON @Natural) (pp ^. hkdMaxCollateralInputsL @era @f))
        ]