+++ libs/cardano-ledger-core/src/Cardano/Ledger/BaseTypes.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
@@ -79,6 +80,10 @@ module Cardano.Ledger.BaseTypes (
 
   -- * Injection
   Inject (..),
+
+  -- * Rationals and Aeson
+  toRationalJSON,
+  parseAsRational,
 )
 where
 
@@ -145,7 +150,7 @@ import Data.Aeson (
   (.:),
   (.=),
  )
-import Data.Aeson.Types (Pair)
+import Data.Aeson.Types (Pair, Parser)
 import qualified Data.Binary.Put as B
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as BSL
@@ -337,11 +342,8 @@ instance
       Just u -> pure u
 
 instance Bounded (BoundedRatio b Word64) => ToJSON (BoundedRatio b Word64) where
-  toJSON = toRationalJSON . unboundRational
-    where
-      toRationalJSON r = case fromRationalRepetendLimited maxDecimalsWord64 r of
-        Right (s, Nothing) -> toJSON s
-        _ -> toJSON r
+  toJSON :: BoundedRatio b Word64 -> Value
+  toJSON = toRationalJSON
 
 instance Bounded (BoundedRatio b Word64) => FromJSON (BoundedRatio b Word64) where
   parseJSON = \case
@@ -855,3 +857,21 @@ class Inject t s where
 -- | Helper function for a common pattern of creating objects
 kindObject :: Text -> [Pair] -> Value
 kindObject name obj = object $ ("kind" .= name) : obj
+
+-- | Loseless way of representing BoundedRational as JSON that uses decimals when feasible
+toRationalJSON :: BoundedRational a => a -> Value
+toRationalJSON br =
+  case fromRationalRepetendLimited 20 r of
+    Right (s, Nothing) -> toJSON s
+    _ -> toJSON r
+  where
+    r = unboundRational br
+
+-- | Aeson parser that works with the result of toRationalJSON
+-- | (It adapts a BoundedRational so that it uses the Rational parser instead)
+parseAsRational :: BoundedRational r => Parser Rational -> Parser r
+parseAsRational parser = do
+  rational <- parser
+  case boundRational rational of
+    Nothing -> fail "Cannot convert to BoundRational"
+    Just x -> return x