+++ libs/cardano-ledger-core/src/Cardano/Ledger/Plutus/ExUnits.hs
@@ -27,7 +27,7 @@ module Cardano.Ledger.Plutus.ExUnits (
 where
 
 import Cardano.Ledger.BaseTypes (
-  BoundedRational (unboundRational),
+  BoundedRational (boundRational, unboundRational),
   NonNegativeInterval,
  )
 import Cardano.Ledger.Binary (
@@ -51,6 +51,7 @@ import Control.Monad (when)
 import Data.Aeson (
   FromJSON (..),
   ToJSON (..),
+  Value,
   object,
   withObject,
   (.:),
@@ -59,6 +60,7 @@ import Data.Aeson (
 import Data.DerivingVia (InstantiatedAt (..))
 import Data.Int (Int64)
 import Data.Measure (BoundedMeasure, Measure)
+import Data.Scientific (fromRationalRepetendLimited)
 import Data.Semigroup (All (..))
 import Data.Word (Word64)
 import GHC.Generics (Generic)
@@ -108,14 +110,14 @@ instance ToJSON ExUnits where
   toJSON exUnits@(ExUnits _ _) =
     let ExUnits {exUnitsMem, exUnitsSteps} = exUnits
      in object
-          [ "exUnitsMem" .= toJSON exUnitsMem
-          , "exUnitsSteps" .= toJSON exUnitsSteps
+          [ "memory" .= toJSON exUnitsMem
+          , "steps" .= toJSON exUnitsSteps
           ]
 
 instance FromJSON ExUnits where
   parseJSON = withObject "exUnits" $ \o -> do
-    exUnitsMem <- checkWord64Bounds =<< o .: "exUnitsMem"
-    exUnitsSteps <- checkWord64Bounds =<< o .: "exUnitsSteps"
+    exUnitsMem <- checkWord64Bounds =<< o .: "memory"
+    exUnitsSteps <- checkWord64Bounds =<< o .: "steps"
     pure $ ExUnits {exUnitsMem, exUnitsSteps}
     where
       checkWord64Bounds n =
@@ -163,9 +165,34 @@ instance NoThunks Prices
 
 instance NFData Prices
 
-instance ToJSON Prices
+instance ToJSON Prices where
+  toJSON Prices {prSteps, prMem} =
+    -- We cannot round-trip via NonNegativeInterval, so we go via Rational
+    object
+      [ "priceSteps" .= toRationalJSON (unboundRational prSteps)
+      , "priceMemory" .= toRationalJSON (unboundRational prMem)
+      ]
+    where
+      toRationalJSON :: Rational -> Value
+      toRationalJSON r =
+        case fromRationalRepetendLimited 20 r of
+          Right (s, Nothing) -> toJSON s
+          _ -> toJSON r
 
-instance FromJSON Prices
+instance FromJSON Prices where
+  parseJSON =
+    withObject "prices" $ \o -> do
+      steps <- o .: "priceSteps"
+      mem <- o .: "priceMemory"
+      prSteps <- checkBoundedRational steps
+      prMem <- checkBoundedRational mem
+      return Prices {prSteps, prMem}
+    where
+      -- We cannot round-trip via NonNegativeInterval, so we go via Rational
+      checkBoundedRational r =
+        case boundRational r of
+          Nothing -> fail ("too much precision for bounded rational: " ++ show r)
+          Just s -> return s
 
 -- | Compute the cost of a script based upon prices and the number of execution
 -- units.